Create Lexer.py
Contains class definitions for Tokens and Token templates. Tokens are pretty much "words" that the parser must divide the input string into before parsing it. Token templates contain a name, a regular expression for recognizing specific string patterns in the input and a function that processes the recognized string before storing the information inside a new Token object. The "lexer" function takes a list of token templates and an input string and performs lexical analysis. The input is split into a list of tokens defined by the token templates and that list is returned and fed into the parser.