Create Parser.py
Functions for creating custom context-free grammars and churning out abstract syntax trees out of those grammars. A grammar rule consists of a stringified rewrite rule ("EXP int") that uses predefined tokens as terminal symbols; a name for the rule; a custom function that takes the list of tokens consumed by the rule as input and returns how they'll appear in the final abstract syntax tree; and a list of "anti-lookahead" symbols before which the rule cannot be applied, for setting precedence. The parse function takes a dictionary of rules as the grammar, empty dictionary as chart, a list of tokens from the lexer, and a starting rule of the grammar that the entire token list should parse into. The function uses a modified version of Earley's parsing algorithm to return a custom abstract syntax tree if the parse is successful. Additional syntactical explanations are provided at the top of the file and example code is provided at the bottom.